It looks like you have removed the `setup` feature and its associated files, as well as updated some other features in your repository. Here's a summary of the changes:
1. The `setup` feature and its associated files (`README.md`, `devcontainer-feature.json`, and `install.sh`) have been removed.
2. In the `tfenv` feature, the `dependsOn` property now includes the `container-dotfiles` feature instead of `common-utils`. The version number has also been incremented to `0.0.8`.
3. In the `update-conda` feature, the `instances` property value for the `miniforge` feature has been changed to just `miniforge`, and the version number has been incremented to `0.0.6`.

It's important to note that removing a feature entirely may have unintended consequences depending on how it is being used in other parts of your project or workflow. If you intended to keep some aspect of the `setup` feature, you may need to extract and adapt its functionality to another feature or script. Similarly, if you've made changes to other features that depend on the removal of `setup`, you may need to update those as well.

Overall, these changes appear to be relatively minor, but it's always a good idea to double-check that your project continues to function correctly after making modifications like this.

 update devcontainer-feature.json for vimrc: change installsAfter order and increment version number to 0.0.6.