commit(src/powershell-modules): update devcontainer-feature.json and install.sh scripts to support PowerShell modules array and improve installation process
update devcontainer-feature.json file to version 0.0.17 and change "version" field to a string type
update install.sh script to handle multiple PowerShell modules by using an array and looping through it for installation.

This commit improves the handling of PowerShell modules in the container by allowing users to specify multiple modules to be installed during container creation. It also updates the devcontainer-feature.json file to support a newer version number and proper data type for the "version" field.