commit (src/powershell-modules): update devcontainer-feature.json and install.sh files to use 'Set-PSRepository' with the PSGallery source location and version 0.0.65
In this commit, I updated the devcontainer-feature.json file to set the version number to 0.0.65 and changed the onCreateCommand for powershell-modules in the install.sh script to use 'Set-PSRepository' with the PSGallery source location instead of using 'sudo -- sh -c'. This change is intended to improve the security and reliability of the PowerShell module installation process.