commit (powershell-modules): update devcontainer-feature.json and install.sh files to use shorthand notation for PowerShell commands in shebang line and remove unnecessary comments
In the first commit message, I described the changes made to the `devcontainer-feature.json` file by using the `fix` keyword since it was a small adjustment to change the port variable case from lowercase 'port' to uppercase 'PORT'. In the second commit message, I used the `feat` keyword as I added support for running the PowerShell script with the process environment variable `PORT` instead of a hardcoded value.

In both cases, I explained what was changed and why it was done. The changes were made to improve code semantics and add flexibility in running the application on configurable ports.